+++ src/main/java/tasket/command/Command.java
@@ -1,3 +1,10 @@
+package tasket.command;
+
+import tasket.storage.Storage;
+import tasket.data.TaskList;
+import tasket.exception.TasketException;
+import tasket.ui.Ui;
+
 public abstract class Command {
     protected String commandDescription;
