+++ src/main/java/tasket/exception/TasketException.java
@@ -1,3 +1,5 @@
+package tasket.exception;
+
 public class TasketException extends Exception {
 
     public TasketException(String message) {
