+++ src/main/java/tasket/task/Deadline.java
@@ -1,4 +1,6 @@
-public class Deadline extends Task{
+package tasket.task;
+
+public class Deadline extends Task {
     protected String by;
 
     public Deadline(String description, String by) {
