+++ src/main/java/tasket/task/Event.java
@@ -1,3 +1,5 @@
+package tasket.task;
+
 public class Event extends Task {
     protected String from;
     protected String to;
