+++ src/main/java/tasket/task/Task.java
@@ -1,4 +1,6 @@
-public class Task {
+package tasket.task;
+
+public abstract class Task {
     protected String description;
     protected boolean isDone;
