+++ src/main/java/tasket/task/ToDo.java
@@ -1,3 +1,5 @@
+package tasket.task;
+
 public class ToDo extends Task {
 
     public ToDo(String description) {
