+++ src/main/java/tasket/ui/Ui.java
@@ -1,3 +1,5 @@
+package tasket.ui;
+
 import java.util.Scanner;
 
 public class Ui {
